Added a type string validity check
cosmic went into an infinite loop when I compiling some code, so I traced it to this and added an error message. I also added two assertions to other functions.

basically, when rewriting the position of the keyword volatile/const, cosmic didn't check if the position it calculated was identical to the current position. such a situation would only happen in an invalid type definition, so an error message was added if this occurred.