Added functionality to parse types with multiple identifiers separated by commas
For all types other than bitfields, these changes should work. There is one thing that might be able to silently go wrong:

struct {int a;long b;} j,k;

^ that declaration will work as expected,  but if there was a way to trick the parser into defining different struct/union/enum for j and k then it would ignore the definition for k and use the one it made for j. I don't think that could happen, but anyway.

Also, any warning that would be given while parsing a struct/union/enum definition will probably be given once for each identifier it is being declared for when in the same declaration statement. That is a side effect from the way I am parsing the type in that situation.