Bugfixes + add basic profiler to simulator
attemptStackMemoryAccessReduction() was applying an optimization that wasn't always valid, this has been fixed. A more complicated detection for validity would be nice to make because that particular optimization is usually valid (but not always). Basically, a memory read to an unknown location between two memory writes to the same location does not mean that the first memory write can be removed because the memory read could read from the location that was just written.

Also, a bug was fixed with messages not printing correctly when in the simulator. It was due to printf specifiers not being the correct type for a system with 16bit int.

A profiler to the simulator was added. It also prints the current call stack in real time to the console.